@@ -1,5 +1,4 @@
-using Microsoft.AspNetCore.Hosting.Internal;
-using Microsoft.Extensions.Hosting;
+using Microsoft.AspNetCore.Hosting;
 using NBitcoin;
 using NBitcoin.RPC;
 using System;
@@ -19,10 +18,10 @@ namespace WalletWasabi.Backend
 		public WebsiteTorifier WebsiteTorifier { get; }
 		public Global Global { get; }
 
-		public InitConfigStartupTask(Global global, IHostingEnvironment hostingEnvironment)
+		public InitConfigStartupTask(Global global, IWebHostEnvironment hostingEnvironment)
 		{
 			Global = global;
-			WebsiteTorifier = new WebsiteTorifier(((HostingEnvironment)hostingEnvironment).WebRootPath);
+			WebsiteTorifier = new WebsiteTorifier(hostingEnvironment.WebRootPath);
 		}
 
 		public async Task ExecuteAsync(CancellationToken cancellationToken)
