@@ -3,6 +3,7 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc.ModelBinding;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using NBitcoin;
 using Swashbuckle.AspNetCore.Swagger;
@@ -63,7 +64,7 @@ namespace WalletWasabi.Backend
 		// This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
 #pragma warning disable IDE0060 // Remove unused parameter
 
-		public void Configure(IApplicationBuilder app, IHostingEnvironment env, Global global)
+		public void Configure(IApplicationBuilder app, IWebHostEnvironment env, Global global)
 #pragma warning restore IDE0060 // Remove unused parameter
 		{
 			app.UseStaticFiles();
@@ -81,9 +82,7 @@ namespace WalletWasabi.Backend
 
 			app.UseResponseCompression();
 
-			app.UseMvc();
-
-			var applicationLifetime = app.ApplicationServices.GetRequiredService<IApplicationLifetime>();
+			var applicationLifetime = app.ApplicationServices.GetRequiredService<IHostApplicationLifetime>();
 			applicationLifetime.ApplicationStopping.Register(() => OnShutdown(global)); // Don't register async, that won't hold up the shutdown
 		}
 
